chore: Update .gitignore and index.html
The .gitignore file was updated to include the `script.py` file. This file was previously not being tracked by Git.

In the index.html file, the interval for calling the `predictImage` function was changed from 2000 milliseconds to 1000 milliseconds. This will result in more frequent predictions being displayed on the webpage.

These changes were made to improve the development workflow and enhance the user experience on the website.