chore: Update capture_images_from_camera function
Refactor the capture_images_from_camera function in script.py to add a pause functionality. This allows the user to pause and resume the image capturing process by pressing the 'p' key. The pause state is indicated with console messages. This change improves the user experience and provides more control over the image capturing process.